Need to finish writing demo_wheel.py
tutorial.py will produce an SVG that acts as a tutorial for how
demo_wheel.py works.

demo_wheel.py will act as a standalone demonstration of how to use the
component library API to draw a barcode wheel.

First, demo_wheel.py needs to be finished, and then factored out into a
component library, so that the component library can be used by a cli
app, a web app, etc.

Seperating the inner workings of generating barcode wheels from the User
Interface should help when work on the web frontend begins, as the
boundary will be clear.

Signed-off-by: Matthew Willcockson <matthew@willcockson.us>