Placeholders are generated correctly
I was worried that my model of creating wheels with empty spaces where
text and pictures would go wouldn't work, but it turns out mozman is THE
MAN and wrote the svgwrite library to hold object references for
svgwrite.Drawing.add() calls, which means if YOU hold a reference to the
object after creating and adding it, you can still modify its
attributes, like the text.

I hope this works for hrefs in images.

If not, I can probably always just make the wheel after parsing
everything, filling in text and such as I go along drawing the wheel.

Now have a rough framework for inserting placeholders

Also, still need to figure out how to insert text and scale it inside
the given box.

In the original OG first iteration of this project, I was trying to
scale each slice to the content, and wrap the names around the circle
and everything.

For these to be useable, they need to be consistent.

In order to perform fast lookups, the information needs to be located in
the same place, and in the same style, everywhere.

This means I don't have to mess with text scaling anyways, or sizing
each slice to the content, as the most effective design will have all
slices equally sized, and each part having the same font size, with the
text start positions all aligned to some common reference point.

Also, there's a looming worry about how performant this code will be,
and if I should just output the SVG to the brower directly, and maybe
should I just redo this entire thing in javascript and do it all client
side???

Also, I'm almost definitely going to require pictures to be uploaded
using URLs, as opposed to receiving pictures directly, unless protecting
against maliciously large uploads, JPEG-processor attacks, etc is a
feasible thing.

I could also create a login system, so that only users I allow can use
the site.

This would solve the worry around file-uploads (except size) and caching
filled out forms.

Really need to break those equations for calculating various things into
their own, self-contained helper functions, or figure out a better way
to reuse those computations in the flow of everything, so that it's
easier to follow what the program is doing.

Signed-off-by: Matthew Willcockson <matthew@willcockson.us>