Text fits in box.
file: test_groups.py

Limitations:
 - text must all use the same shaping model
 - all glyphs must be available in the font specified

So HarfBuzz can only handle text runs that are the same language, and it
cannot handle bidirectional text (e.g. mixed Arabic and English).

This text must be preprocessed and fed to HarfBuzz piece by piece.

Also, with fontconfig sourcing font files for me, I'm basically trying
to do the legwork the browser is doing when it finds the font files that
it can use to render the glyphs it needs to in order to render the text.

The browser is smart about finding font files that have the glyphs it
needs.

My use of fontconfig does not.

This works fine for Latin text, and DejaVu Sans has enough glyphs that I
think it'll be able to handle most things thrown at it.

The edge cases are now:

 - Use fribidi to preprocess bidirectional text into runs.
 - Further process each run to find fonts with glyphs for each character
 - Split text on these font boundaries
 - Split these runs on shaping model boundaries
 - Then pass each of THOSE runs to HarfBuzz for shaping
 - Use HarfBuzz to get glyph shape and layout information
 - Then turn the glyph outlines into SVGs

This STILL has the limitation that HarfBuzz has: it only handles text on
one long, imaginary line.

However, this is enough to turn it into a web services, where a WOFF
(Web Font) to OpenType converter is used to take a WOFF and a string of
text, and return an SVG that can be used as an image with text that will
natively fill the region specified in the HTML <img> tag!

This could be a nice web service with an API and everything!

ICU seems to be more integrated project for this.

https://catherineh.github.io/programming/2018/02/01/text-to-svg-paths

The above link give a way to turn glyph information into outlines, and
combined with kerning and positioning information, could be used to
render text as an SVG, which can be scaled very easily.

Bounding box calculation may still be incorrect:

text_bounding_box() doesn't take into account descenders or ascenders,
or offsets of the characters: a 'y' may not be taller than a 'T', but
the descender would make 'yT' have an overall taller bounding box than
just 'T'.

Also, SVG uses font-size in user coordinates, meaning if the enclosing
<rect> is 100x100, font-size="100" will work, except that the font will
be baseline to baseline, as pointed out by the internet, when googling
"svg font-size".

Line height is determined by the rendering engine, I think, but can be
ignored by setting a viewBox that clips that extra space, since the text
size is known.

Using HarfBuzz requires installing packages with platform-dependent
names, and building HarfBuzz from source with custom ./configure options
specified in the GitHub repo for use with pyGObject.

The svgwrite library holds object references for svgwrite.Drawing.add()
calls, which means old references to the same objects can be used modify
their attributes, like the text. This allows the contents of tags to be
updated after adding them to the drawing canvas, meaning a "blank" pie
can be generated, and then multiple passes can be made to fill in
placeholders, generate an output, and then repeat, replacing each
placeholders' contents as you go along.

In the original iteration of this project, I was trying to scale each
slice to the content, and wrap the names around the edge of the pie.

For the generated barcode wheels to be useable, they need to be
consistent.

In order to perform fast lookups, the information needs to be located in
the same place, and in the same style, everywhere.

This most likely means that things like scaling the text to fit the
predifined placeholder box, or sizing each slice to its content, will be
pointless anyways, as the most effective design will have all slices
equally sized, and each part having the same font size, with the text
start positions all aligned to some common reference point.

Pictures will probably be required to be uploaded using URLs, as opposed
to receiving pictures directly, unless protecting against maliciously
large uploads, JPEG-processor attacks, etc is feasible.

A login system could be used to mitigate this risk, so that only allowed
users can use the site.

This would solve the worry around file-uploads (except size) and caching
filled out forms.

Signed-off-by: Matthew Willcockson <matthew@willcockson.us>