Text placement works, and CSVs can be parsed
Now beginning work on reworking the way the wheel is generated:

- Find the name or PLU that will require the most space (i.e. is the
longest). Thise text can only be a minimum font size before it is
unreadable
- Use that to determine how many slices are in each wheel
- Make the outline
- Fill each slice with a placeholder group, rotating each with a <use>
into place
- Split dataset into sections for each wheel
- Run through each section of the dataset, filling in a complete wheel,
as we go, filling in blanks for any unfilled slices in the last section,
saving the wheel to a new page file, then moving to the next section

text_filled_area() WORKS!

Mostly: There's still a little bit of the last character being cut off,
but it's close enough that I'm happy.

The two files I used to achieve this were:

- single_char.py

This takes a single character (by default M if none is given) and just
puts it in an SVG <text> tag inserted at (0, 0) and with the font size
set to match the units returned by glyph_metrics()

Then, you can set the x and y positions that you want vertical and
horizontal lines drawn at, so that you can pull it up in an SVG viewer,
like Inkscape, to see and measure where everything is.

I used this and trial and error to figure out the calculations done in
the current text_bounding_box()

- harfuzz_metrics.py

Was reworked to show how to use the text_filled_area(), and to let me
fine-tune some of the edge cases for text_bounding_box() It draws an
orange background to show what the <use> tag's area is, then a gray box
to show the <text> bounding box.  The grey box should always be inside
the orange box, and should be the same as the orange box in at least one
dimension (i.e. same width or same height, or both).  The text should
stay inside the grey box, and should be completely visible, and should
match the gray box in every dimension.

This seems to hold, as long as the unicode string being requested exists
completely within the font file that fontconfig finds for "sans-serif".

The file no_text_filled_area.py shows how to replace text_filled_area()
with calculations of the font size and text placement to mimic where the
renderer will place the <symbol>.

This functionality hasn't been wrapped up, yet, in a function that
replaces text_filled_area(), but that function will return an
svgwrite.text.Text with all the font_size and insert set properly.

No more adding to <defs>!

Also, the function font_size_in_box() gives the font size of the given
text if it were to fill the given area. This will be used for
determining if auto-scaled text in a box is too small, by comparing it
to a predetermined "smallest readable font-size".

Ideally, text_bounding_box().width/height are the only things necessary, and may
even be split off into their own, speedier(?) function, as those will be
used for the initial demo_wheel.py completion, to ensure text doesn't
overflow it's assigned area.

Another function will have to be written to accommodate vertically
centering text that hasn't had any font-size adjustments, to place the
text string in the center of the given box, as well as horizontally
centered.

Signed-off-by: Matthew Willcockson <matthew@willcockson.us>